# openapi_preview.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Script Rating API",
        "description": "API for rating and analyzing scripts using clean architecture",
        "version": "0.1.0"
    },
    "paths": {
...
        "/api/rag/query": {
            "post": {
                "tags": [
                    "RAG"
                ],
                "summary": "Query RAG system",
                "description": "Query the RAG system for relevant legal and reference content using vector search.",
                "operationId": "query_rag_api_rag_query_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/RAGQueryRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/RAGQueryResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/rag/corpus/update": {
            "post": {
                "tags": [
                    "RAG"
                ],
                "summary": "Update corpus",
                "description": "Add new content to the RAG knowledge corpus.",
                "operationId": "update_corpus_api_rag_corpus_update_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CorpusUpdateRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CorpusUpdateResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/rag/corpus": {
            "get": {
                "tags": [
                    "RAG"
                ],
                "summary": "List corpus documents",
                "description": "Get a list of documents in the RAG corpus.",
                "operationId": "list_corpus_api_rag_corpus_get",
                "parameters": [
                    {
                        "name": "category",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "$ref": "#/components/schemas/Category"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "description": "Filter by category",
                            "title": "Category"
                        },
                        "description": "Filter by category"
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "maximum": 200,
                            "minimum": 1,
                            "description": "Maximum number of documents to return",
                            "default": 50,
                            "title": "Limit"
                        },
                        "description": "Maximum number of documents to return"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/rag/corpus/{document_id}": {
            "get": {
                "tags": [
                    "RAG"
                ],
                "summary": "Get corpus document",
                "description": "Get detailed information about a specific corpus document.",
                "operationId": "get_corpus_document_api_rag_corpus__document_id__get",
                "parameters": [
                    {
                        "name": "document_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Document Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "RAG"
                ],
                "summary": "Delete corpus document",
                "description": "Remove a document from the RAG corpus.",
                "operationId": "delete_corpus_document_api_rag_corpus__document_id__delete",
                "parameters": [
                    {
                        "name": "document_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Document Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/rag/queries": {
            "get": {
                "tags": [
                    "RAG"
                ],
                "summary": "List RAG queries",
                "description": "Get a list of recent RAG queries and their results.",
                "operationId": "list_rag_queries_api_rag_queries_get",
                "parameters": [
                    {
                        "name": "limit",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "maximum": 100,
                            "minimum": 1,
                            "description": "Maximum number of queries to return",
                            "default": 20,
                            "title": "Limit"
                        },
                        "description": "Maximum number of queries to return"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Category": {
                "type": "string",
                "enum": [
                    "violence",
                    "sexual_content",
                    "language",
                    "alcohol_drugs",
                    "disturbing_scenes"
                ],
                "title": "Category"
            },
...
            "CorpusUpdateRequest": {
                "properties": {
                    "content": {
                        "type": "string",
                        "title": "Content",
                        "description": "New content to add"
                    },
                    "category": {
                        "$ref": "#/components/schemas/Category",
                        "description": "Content category"
                    },
                    "source_title": {
                        "type": "string",
                        "title": "Source Title",
                        "description": "Source title"
                    },
                    "source_metadata": {
                        "anyOf": [
                            {
                                "additionalProperties": true,
                                "type": "object"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Source Metadata",
                        "description": "Additional metadata"
                    }
                },
                "type": "object",
                "required": [
                    "content",
                    "category",
                    "source_title"
                ],
                "title": "CorpusUpdateRequest",
                "description": "Request model for corpus updates."
            },
            "CorpusUpdateResponse": {
                "properties": {
                    "update_id": {
                        "type": "string",
                        "title": "Update Id",
                        "description": "Unique identifier for the update"
                    },
                    "content_hash": {
                        "type": "string",
                        "title": "Content Hash",
                        "description": "Hash of the added content"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "title": "Updated At",
                        "description": "Update timestamp"
                    }
                },
                "type": "object",
                "required": [
                    "update_id",
                    "content_hash",
                    "updated_at"
                ],
                "title": "CorpusUpdateResponse",
                "description": "Response model for corpus updates."
            },
...
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
...
            "RAGQueryRequest": {
                "properties": {
                    "query": {
                        "type": "string",
                        "title": "Query",
                        "description": "Search query"
                    },
                    "category": {
                        "anyOf": [
                            {
                                "$ref": "#/components/schemas/Category"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "description": "Filter by category"
                    },
                    "top_k": {
                        "type": "integer",
                        "title": "Top K",
                        "description": "Number of results to return",
                        "default": 5
                    }
                },
                "type": "object",
                "required": [
                    "query"
                ],
                "title": "RAGQueryRequest",
                "description": "Request model for RAG queries."
            },
            "RAGQueryResponse": {
                "properties": {
                    "query": {
                        "type": "string",
                        "title": "Query",
                        "description": "Original query"
                    },
                    "results": {
                        "items": {
                            "$ref": "#/components/schemas/RAGResult"
                        },
                        "type": "array",
                        "title": "Results",
                        "description": "Retrieved results"
                    },
                    "total_found": {
                        "type": "integer",
                        "title": "Total Found",
                        "description": "Total relevant documents found"
                    }
                },
                "type": "object",
                "required": [
                    "query",
                    "results",
                    "total_found"
                ],
                "title": "RAGQueryResponse",
                "description": "Response model for RAG queries."
            },
...
        }
    }
}